restructure utilities
